+++ hardware_interface/src/resource_manager.cpp
@@ -583,7 +583,7 @@ public:
     switch (target_state.id())
     {
       case State::PRIMARY_STATE_UNCONFIGURED:
-        switch (hardware.get_lifecycle_state().id())
+        switch (hardware.get_lifecycle_id())
         {
           case State::PRIMARY_STATE_UNCONFIGURED:
             result = true;
@@ -607,7 +607,7 @@ public:
         }
         break;
       case State::PRIMARY_STATE_INACTIVE:
-        switch (hardware.get_lifecycle_state().id())
+        switch (hardware.get_lifecycle_id())
         {
           case State::PRIMARY_STATE_UNCONFIGURED:
             result = configure_hardware(hardware);
@@ -627,7 +627,7 @@ public:
         }
         break;
       case State::PRIMARY_STATE_ACTIVE:
-        switch (hardware.get_lifecycle_state().id())
+        switch (hardware.get_lifecycle_id())
         {
           case State::PRIMARY_STATE_UNCONFIGURED:
             result = configure_hardware(hardware);
@@ -651,7 +651,7 @@ public:
         }
         break;
       case State::PRIMARY_STATE_FINALIZED:
-        switch (hardware.get_lifecycle_state().id())
+        switch (hardware.get_lifecycle_id())
         {
           case State::PRIMARY_STATE_UNCONFIGURED:
             result = shutdown_hardware(hardware);
@@ -1951,8 +1951,8 @@ ResourceManager::get_components_status()
   {
     for (auto & component : container)
     {
-      resource_storage_->hardware_info_map_[component.get_name()].state =
-        component.get_lifecycle_state();
+      resource_storage_->hardware_info_map_[component.get_name()].state = rclcpp_lifecycle::State(
+        component.get_lifecycle_state().id(), component.get_lifecycle_state().label());
     }
   };
 
@@ -2061,8 +2061,7 @@ bool ResourceManager::prepare_command_mode_switch(
       }
       if (
         !start_interfaces_buffer.empty() &&
-        component.get_lifecycle_state().id() ==
-          lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE &&
+        component.get_lifecycle_id() == lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE &&
         !allow_controller_activation_with_inactive_hardware)
       {
         RCLCPP_WARN(
@@ -2072,9 +2071,8 @@ bool ResourceManager::prepare_command_mode_switch(
         return false;
       }
       if (
-        component.get_lifecycle_state().id() ==
-          lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE ||
-        component.get_lifecycle_state().id() == lifecycle_msgs::msg::State::PRIMARY_STATE_ACTIVE)
+        component.get_lifecycle_id() == lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE ||
+        component.get_lifecycle_id() == lifecycle_msgs::msg::State::PRIMARY_STATE_ACTIVE)
       {
         try
         {
@@ -2168,8 +2166,7 @@ bool ResourceManager::perform_command_mode_switch(
       }
       if (
         !start_interfaces_buffer.empty() &&
-        component.get_lifecycle_state().id() ==
-          lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE &&
+        component.get_lifecycle_id() == lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE &&
         !allow_controller_activation_with_inactive_hardware)
       {
         RCLCPP_WARN(
@@ -2179,9 +2176,8 @@ bool ResourceManager::perform_command_mode_switch(
         return false;
       }
       if (
-        component.get_lifecycle_state().id() ==
-          lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE ||
-        component.get_lifecycle_state().id() == lifecycle_msgs::msg::State::PRIMARY_STATE_ACTIVE)
+        component.get_lifecycle_id() == lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE ||
+        component.get_lifecycle_id() == lifecycle_msgs::msg::State::PRIMARY_STATE_ACTIVE)
       {
         try
         {